chore: merge VinePreviewScreenPure files and fix tests
