+++ mobile/test/screens/pure/vine_preview_screen_pure_test.dart
@@ -1,0 +1,263 @@
+// ABOUTME: TDD tests for VinePreviewScreenPure widget
+// ABOUTME: Tests save draft functionality and video preview display behavior
+
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:openvine/models/vine_draft.dart';
+import 'package:openvine/screens/pure/vine_preview_screen_pure.dart';
+import 'package:openvine/services/draft_storage_service.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:video_player/video_player.dart';
+
+void main() {
+  group('VinePreviewScreenPure', () {
+      late DraftStorageService draftService;
+  late VineDraft testDraft;
+
+  setUp(() async {
+    SharedPreferences.setMockInitialValues({});
+    final prefs = await SharedPreferences.getInstance();
+    draftService = DraftStorageService(prefs);
+
+    // Create and save a default test draft
+    testDraft = VineDraft.create(
+      videoFile: File('/path/to/test/video.mp4'),
+      title: '',
+      description: '',
+      hashtags: [],
+      frameCount: 30,
+      selectedApproach: 'hybrid',
+    );
+    await draftService.saveDraft(testDraft);
+  });
+
+
+  group('save draft', () {
+    testWidgets('should have a Save Draft button in app bar', (tester) async {
+      await tester.pumpWidget(
+        ProviderScope(
+          child: MaterialApp(
+            home: VinePreviewScreenPure(draftId: testDraft.id),
+          ),
+        ),
+      );
+
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 100));
+
+      expect(find.text('Save Draft'), findsOneWidget);
+    });
+
+    testWidgets('should save draft when Save Draft button is tapped', (
+      tester,
+    ) async {
+      await tester.pumpWidget(
+        ProviderScope(
+          child: MaterialApp(
+            home: VinePreviewScreenPure(draftId: testDraft.id),
+          ),
+        ),
+      );
+
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 100));
+
+      // Enter metadata - find TextFields by their order (title is first)
+      final textFields = find.byType(TextField);
+      await tester.enterText(textFields.first, 'Test Vine Title');
+
+      // Tap Save Draft
+      await tester.tap(find.text('Save Draft'));
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 100));
+
+      // Verify draft was saved to storage
+      final drafts = await draftService.getAllDrafts();
+      expect(drafts.length, 1);
+      expect(drafts.first.title, 'Test Vine Title');
+      expect(drafts.first.frameCount, 30);
+      expect(drafts.first.selectedApproach, 'hybrid');
+    });
+
+    testWidgets('should save draft with empty fields', (tester) async {
+      await tester.pumpWidget(
+        ProviderScope(
+          child: MaterialApp(
+            home: VinePreviewScreenPure(draftId: testDraft.id),
+          ),
+        ),
+      );
+
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 100));
+
+      // Don't enter any metadata, just save
+      await tester.tap(find.text('Save Draft'));
+      await tester.pump();
+
+      // Verify draft was saved with empty fields
+      final drafts = await draftService.getAllDrafts();
+      expect(drafts.length, 1);
+      expect(drafts.first.title, '');
+      expect(drafts.first.description, '');
+    });
+
+    testWidgets('should not disable Save Draft button when uploading', (
+      tester,
+    ) async {
+      await tester.pumpWidget(
+        ProviderScope(
+          child: MaterialApp(
+            home: VinePreviewScreenPure(draftId: testDraft.id),
+          ),
+        ),
+      );
+
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 100));
+
+      // Save Draft button should always be enabled (independent of upload state)
+      final saveDraftButton = find.text('Save Draft');
+      expect(saveDraftButton, findsOneWidget);
+
+      // Verify it's a TextButton and not disabled
+      final textButton = tester.widget<TextButton>(
+        find.ancestor(of: saveDraftButton, matching: find.byType(TextButton)),
+      );
+      expect(textButton.onPressed, isNotNull);
+    });
+  });
+
+  group('video preview', () {
+    testWidgets(
+      'should show either VideoPlayer or placeholder after initialization',
+      (tester) async {
+        await tester.pumpWidget(
+          ProviderScope(
+            child: MaterialApp(
+              home: VinePreviewScreenPure(draftId: testDraft.id),
+            ),
+          ),
+        );
+
+        // Pump once to trigger initState
+        await tester.pump();
+
+        // Wait for async initialization to complete
+        await tester.pump(const Duration(seconds: 3));
+
+        // After initialization attempt, widget should show either:
+        // - VideoPlayer (if file was valid and initialization succeeded)
+        // - Placeholder icon (if file was invalid or initialization failed)
+        final hasVideoPlayer = find.byType(VideoPlayer).evaluate().isNotEmpty;
+        final hasPlaceholder = find
+            .byIcon(Icons.play_circle_filled)
+            .evaluate()
+            .isNotEmpty;
+
+        // One of these should be true
+        expect(
+          hasVideoPlayer || hasPlaceholder,
+          isTrue,
+          reason:
+              'Should show either VideoPlayer widget or placeholder icon after initialization',
+        );
+      },
+    );
+
+    testWidgets(
+      'should show VideoPlayer widget when video initializes successfully',
+      (tester) async {
+        await tester.pumpWidget(
+          ProviderScope(
+            child: MaterialApp(
+              home: VinePreviewScreenPure(draftId: testDraft.id),
+            ),
+          ),
+        );
+
+        await tester.pump();
+
+        // If video initialized, VideoPlayer widget should exist
+        // In real usage with valid video file, this would pass
+        // In tests with fake file path, this will show placeholder
+        final videoPlayerFinder = find.byType(VideoPlayer);
+
+        // We expect either VideoPlayer (if initialized) or placeholder (if failed)
+        // This verifies conditional rendering logic exists
+        final hasVideoPlayer = videoPlayerFinder.evaluate().isNotEmpty;
+        final hasPlaceholder = find
+            .byIcon(Icons.play_circle_filled)
+            .evaluate()
+            .isNotEmpty;
+
+        // At least one should be present (either video or placeholder)
+        expect(hasVideoPlayer || hasPlaceholder, isTrue);
+      },
+    );
+
+    testWidgets('should show placeholder when video fails to initialize', (
+      tester,
+    ) async {
+      // Create a draft with nonexistent video for this specific test
+      final badDraft = VineDraft.create(
+        videoFile: File('/nonexistent/video.mp4'),
+        title: '',
+        description: '',
+        hashtags: [],
+        frameCount: 30,
+        selectedApproach: 'hybrid',
+      );
+      await draftService.saveDraft(badDraft);
+
+      await tester.pumpWidget(
+        ProviderScope(
+          child: MaterialApp(
+            home: VinePreviewScreenPure(draftId: badDraft.id),
+          ),
+        ),
+      );
+
+      // Wait for initialization attempt to fail
+      await tester.pump();
+      await tester.pump(const Duration(seconds: 3));
+
+      // Should show placeholder icon when video fails
+      expect(find.byIcon(Icons.play_circle_filled), findsOneWidget);
+
+      // Should NOT show VideoPlayer widget when failed
+      expect(find.byType(VideoPlayer), findsNothing);
+    });
+
+    testWidgets('should dispose cleanly without crashes', (tester) async {
+      await tester.pumpWidget(
+        ProviderScope(
+          child: MaterialApp(
+            home: VinePreviewScreenPure(draftId: testDraft.id),
+          ),
+        ),
+      );
+
+      await tester.pump();
+
+      // Navigate away to trigger dispose
+      await tester.pumpWidget(
+        const ProviderScope(
+          child: MaterialApp(home: Scaffold(body: Text('Other screen'))),
+        ),
+      );
+
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 100));
+
+      // Verify dispose completed without crashes
+      expect(find.text('Other screen'), findsOneWidget);
+    });
+  });
+
+  });
+
+}
