--- mobile/test/screens/vine_preview_screen_video_preview_test.dart
@@ -1,153 +1,0 @@
-// ABOUTME: TDD test for video preview functionality in VinePreviewScreenPure
-// ABOUTME: Verifies video controller creation and preview display using state inspection
-
-import 'dart:io';
-
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:openvine/screens/pure/vine_preview_screen_pure.dart';
-import 'package:video_player/video_player.dart';
-
-void main() {
-  group('VinePreviewScreenPure video preview', () {
-    testWidgets(
-      'should show either VideoPlayer or placeholder after initialization',
-      (tester) async {
-        final videoFile = File('/path/to/test/video.mp4');
-
-        await tester.pumpWidget(
-          ProviderScope(
-            child: MaterialApp(
-              home: VinePreviewScreenPure(
-                videoFile: videoFile,
-                frameCount: 30,
-                selectedApproach: 'hybrid',
-              ),
-            ),
-          ),
-        );
-
-        // Pump once to trigger initState
-        await tester.pump();
-
-        // Wait for async initialization to complete
-        await tester.pump(const Duration(seconds: 3));
-
-        // After initialization attempt, widget should show either:
-        // - VideoPlayer (if file was valid and initialization succeeded)
-        // - Placeholder icon (if file was invalid or initialization failed)
-        final hasVideoPlayer = find.byType(VideoPlayer).evaluate().isNotEmpty;
-        final hasPlaceholder = find
-            .byIcon(Icons.play_circle_filled)
-            .evaluate()
-            .isNotEmpty;
-
-        // One of these should be true
-        expect(
-          hasVideoPlayer || hasPlaceholder,
-          isTrue,
-          reason:
-              'Should show either VideoPlayer widget or placeholder icon after initialization',
-        );
-      },
-    );
-
-    testWidgets(
-      'should show VideoPlayer widget when video initializes successfully',
-      (tester) async {
-        // This test will fail because we can't create real video files in tests
-        // But it documents the expected behavior
-        final videoFile = File('/path/to/test/video.mp4');
-
-        await tester.pumpWidget(
-          ProviderScope(
-            child: MaterialApp(
-              home: VinePreviewScreenPure(
-                videoFile: videoFile,
-                frameCount: 30,
-                selectedApproach: 'hybrid',
-              ),
-            ),
-          ),
-        );
-
-        await tester.pump();
-
-        // If video initialized, VideoPlayer widget should exist
-        // In real usage with valid video file, this would pass
-        // In tests with fake file path, this will show placeholder
-        final videoPlayerFinder = find.byType(VideoPlayer);
-
-        // We expect either VideoPlayer (if initialized) or placeholder (if failed)
-        // This verifies conditional rendering logic exists
-        final hasVideoPlayer = videoPlayerFinder.evaluate().isNotEmpty;
-        final hasPlaceholder = find
-            .byIcon(Icons.play_circle_filled)
-            .evaluate()
-            .isNotEmpty;
-
-        // At least one should be present (either video or placeholder)
-        expect(hasVideoPlayer || hasPlaceholder, isTrue);
-      },
-    );
-
-    testWidgets('should show placeholder when video fails to initialize', (
-      tester,
-    ) async {
-      final videoFile = File('/nonexistent/video.mp4');
-
-      await tester.pumpWidget(
-        ProviderScope(
-          child: MaterialApp(
-            home: VinePreviewScreenPure(
-              videoFile: videoFile,
-              frameCount: 30,
-              selectedApproach: 'hybrid',
-            ),
-          ),
-        ),
-      );
-
-      // Wait for initialization attempt to fail
-      await tester.pump();
-      await tester.pump(const Duration(seconds: 3));
-
-      // Should show placeholder icon when video fails
-      expect(find.byIcon(Icons.play_circle_filled), findsOneWidget);
-
-      // Should NOT show VideoPlayer widget when failed
-      expect(find.byType(VideoPlayer), findsNothing);
-    });
-
-    testWidgets('should dispose cleanly without crashes', (tester) async {
-      final videoFile = File('/path/to/test/video.mp4');
-
-      await tester.pumpWidget(
-        ProviderScope(
-          child: MaterialApp(
-            home: VinePreviewScreenPure(
-              videoFile: videoFile,
-              frameCount: 30,
-              selectedApproach: 'hybrid',
-            ),
-          ),
-        ),
-      );
-
-      await tester.pump();
-
-      // Navigate away to trigger dispose
-      await tester.pumpWidget(
-        const ProviderScope(
-          child: MaterialApp(home: Scaffold(body: Text('Other screen'))),
-        ),
-      );
-
-      await tester.pumpAndSettle();
-
-      // Verify dispose completed without crashes
-      expect(find.text('Other screen'), findsOneWidget);
-    });
-  });
-}
